+++ distant-net/tests/typed_tests.rs
@@ -30,7 +30,7 @@ impl ServerHandler for TestServerHandler {
 async fn should_be_able_to_send_and_receive_typed_payloads_between_client_and_server() {
     let (t1, t2) = InmemoryTransport::pair(100);
 
-    let _ = Server::new()
+    let _server = Server::new()
         .handler(TestServerHandler)
         .verifier(Verifier::none())
         .start(OneshotListener::from_value(t2))